requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pyhd8ed1ab_1
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.0=mkl
box2d-py=2.3.5=pypi_0
brotli=1.1.0=h2466b09_2
brotli-bin=1.1.0=h2466b09_2
brotli-python=1.1.0=py310h9e98ed7_2
bzip2=1.0.8=h2466b09_7
c-ares=1.34.3=h2466b09_1
ca-certificates=2024.9.24=haa95532_0
cairo=1.18.0=h32b962e_3
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py310ha8f682b_0
charset-normalizer=3.4.0=pyhd8ed1ab_0
click=8.1.7=pypi_0
cloudpickle=3.1.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.3.0=py310hc19bc0b_2
cpython=3.10.15=py310hd8ed1ab_2
cuda-cccl=12.6.77=0
cuda-cccl_win-64=12.6.77=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-opencl=12.6.77=0
cuda-opencl-dev=12.6.77=0
cuda-profiler-api=12.6.77=0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
cycler=0.12.1=pyhd8ed1ab_0
cython=0.29.33=pypi_0
debugpy=1.8.7=py310h9e98ed7_0
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
double-conversion=3.3.0=h63175ca_0
etils=1.11.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
fasteners=0.19=pypi_0
filelock=3.16.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h765892d_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.54.1=py310h38315fa_1
freetype=2.12.1=hdaf720e_2
fsspec=2024.12.0=pypi_0
git=2.45.2=haa95532_1
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glfw=2.8.0=pypi_0
graphite2=1.3.13=h63175ca_1003
grpcio=1.67.1=py310h0288bfe_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
h2=4.1.0=pyhd8ed1ab_0
harfbuzz=9.0.0=h2bedf89_1
hpack=4.0.0=pyh9f0ad1d_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=75.1=he0c23c2_0
idna=3.10=pyhd8ed1ab_0
imageio=2.36.0=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.5.0=pyha770c72_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2024.2.1=h57928b3_1083
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.29.0=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh5737063_1
kiwisolver=1.4.7=py310hc19bc0b_0
krb5=1.21.3=hdf4eb48_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libabseil=20240722.0=cxx17_he0c23c2_1
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.1.0=h2466b09_2
libbrotlidec=1.1.0=h2466b09_2
libbrotlienc=1.1.0=h2466b09_2
libcblas=3.9.0=12_win64_mkl
libclang13=19.1.3=default_ha5278ca_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.7.77=0
libcurand-dev=10.3.7.77=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.22=h2466b09_0
libexpat=2.6.4=he0c23c2_0
libffi=3.4.2=h8ffe710_5
libgcc=14.2.0=h1383e82_1
libglib=2.82.2=h7025463_0
libgomp=14.2.0=h1383e82_1
libgrpc=1.67.1=h7aa3b8a_0
libiconv=1.17=hcfcfb64_2
libintl=0.22.5=h5728263_3
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=12_win64_mkl
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvfatbin=12.6.77=0
libnvfatbin-dev=12.6.77=0
libnvjitlink=12.4.127=0
libnvjitlink-dev=12.4.127=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.44=h3ca93ac_0
libprotobuf=5.28.2=hcaed137_0
libre2-11=2024.07.02=h4eb7d71_1
libsodium=1.0.20=hc70643c_0
libsqlite=3.47.0=h2466b09_1
libtiff=4.7.0=hfc51747_1
libuv=1.49.2=h2466b09_0
libwebp-base=1.4.0=hcfcfb64_0
libwinpthread=12.0.0.r4.gg4f2fc60ca=h57928b3_8
libxcb=1.17.0=h0e4246c_0
libxml2=2.13.4=h442d1da_2
libxslt=1.1.39=h3df6e99_0
libzlib=1.3.1=h2466b09_2
markdown=3.6=pyhd8ed1ab_0
markupsafe=3.0.2=py310h38315fa_0
matplotlib=3.9.3=py310h5588dad_0
matplotlib-base=3.9.3=py310h37e0a56_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2021.4.0=h0e2418a_729
moviepy=2.0.0=pypi_0
mpmath=1.3.0=pyhd8ed1ab_0
mujoco=2.2.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.4.2=pyhd8ed1ab_1
numpy=1.23.1=pypi_0
openjpeg=2.5.2=h3d672ee_0
openssl=3.4.0=h2466b09_0
packaging=24.1=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.44=h3d7b363_2
pickleshare=0.7.5=py_1003
pillow=10.4.0=pypi_0
pip=24.3.1=pyh8b19718_0
pixman=0.43.4=h63175ca_0
platformdirs=4.3.6=pyhd8ed1ab_0
proglog=0.1.10=pypi_0
prompt-toolkit=3.0.48=pyha770c72_0
protobuf=5.28.3=pypi_0
psutil=6.1.0=py310ha8f682b_0
pthread-stubs=0.4=h0e40799_1002
pure_eval=0.2.3=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pygame=2.1.0=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyopengl=3.1.9=pypi_0
pyparsing=3.2.0=pyhd8ed1ab_1
pyside6=6.8.0.2=py310h60c6385_0
pysocks=1.7.1=pyh0701188_6
python=3.10.15=hfaddaf0_2_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python_abi=3.10=5_cp310
pytorch=2.3.1=py3.10_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pywin32=307=py310h9e98ed7_3
pyyaml=6.0.2=py310ha8f682b_1
pyzmq=26.2.0=py310h656833d_3
qhull=2020.2=hc790b64_5
qt6-main=6.8.0=hfb098fa_0
re2=2024.07.02=hd3b24a8_1
requests=2.32.3=pyhd8ed1ab_0
sentry-sdk=2.18.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.3.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
swig=4.2.1.post0=pypi_0
sympy=1.13.3=pyh04b8f61_4
tbb=2021.7.0=h91493d7_0
tensorboard=2.18.0=pyhd8ed1ab_1
tensorboard-data-server=0.7.0=py310h5588dad_2
tk=8.6.13=h5226925_1
torchaudio=2.3.1=pypi_0
torchvision=0.18.1=pypi_0
tornado=6.4.1=py310ha8f682b_1
tqdm=4.67.1=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
ucrt=10.0.22621.0=h57928b3_1
unicodedata2=15.1.0=py310ha8f682b_1
urllib3=2.2.3=pyhd8ed1ab_0
vc=14.3=h8a93ad2_22
vc14_runtime=14.40.33810=hcc2c482_22
vs2015_runtime=14.40.33810=h3bf8584_22
wandb=0.18.7=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.1.3=pyhd8ed1ab_1
wheel=0.44.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyh7428d3b_7
xorg-libxau=1.0.11=h0e40799_1
xorg-libxdmcp=1.1.5=h0e40799_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zeromq=4.3.5=ha9f60a1_6
zipp=3.20.2=pyhd8ed1ab_0
zlib=1.3.1=h2466b09_2
zstandard=0.23.0=py310he5e10e1_1
zstd=1.5.6=h0ea2cb4_0
